BUG修复: 无法使用绝对路径运行 Seraphine #172 功能优化: make.ps1添加参数 -dbg, 用于调试时构建, 将不会删除dist下的文件, 也不会生成压缩包;

+++ make.ps1
@@ -1,7 +1,26 @@
+<#
+.SYNOPSIS
+    该脚本用于打包 "Seraphine" 。
+
+.PARAMETER dest
+    输出的目标路径。默认为当前目录。
+
+.PARAMETER dbg
+    是否启用调试模式。如果启用，将不会删除 `.\dist` 目录，也不会创建 zip 文件。
+
+.EXAMPLE
+    .\make.ps1 -dbg
+#>
+
 param(
     [Parameter()]
-    [String]$dest = "."
+    [String]$dest = ".",
+    [Switch]$dbg
 )
+
+if ($dbg -and (Test-Path .\dist)) {
+    rm -r -Force .\dist
+}
 
 pyinstaller -w -i .\app\resource\images\logo.ico main.py
 rm -r -fo .\build
@@ -16,5 +35,8 @@ rm -Path .\dist\Seraphine\app\config* -r
 rm -Path .\dist\Seraphine\app\resource\game* -r
 rm -r .\dist\Seraphine\app\resource\i18n\Seraphine.zh_CN.ts
 rm -r .\dist\Seraphine\app\view
-7z a $dest\Seraphine.zip .\dist\Seraphine\* -r
-rm -r .\dist
+
+if (! $dbg) {
+    7z a $dest\Seraphine.zip .\dist\Seraphine\* -r
+    rm -r .\dist
+}
